Beginnings of cellManager started
Started the beginnings of a background cell phone manager that  inherits from quickCom. This will queue up requests and keep everything in synk. The idea is to get all the cell stuff out of the OS class. It should just be a manager of panels.

Also reset volume to 30. Was at 100, too loud!