Killing old bugs
Decided to use "live text" for status updates. Had to do a bunch of work to that. Sort out the logic. Be able to force last color after the end of a fade out. Always been a problem. Also buggered up a way to erase and reset if a new message comes in before the last is finished.