Phone is at an ok stopping point, starting mp3 stuff.
The cellphone can recieve calls, make calls. It can't ring a call in, yet. But that'll get worked out in time.

I told Adafruit I'd look at their mp3 library and see if it looks ok. I'd done a change in there and they want to roll it in. About a year or so later.. So be it. That's what I'm up to now.