Scroling test fixed.
Turns out I was grabbing wrong pointers for the list and things like that. Having the list be headed by a list object, little tough. YOu have to really keep track of "List item or header?" The big issue was asking for the "firstItem() and, whoops that's the header in this case. What I needed to call was getTailObj(0); Live and learn. I need to add a call getTail(). Maybe later. Maybe its already there?