mp3 player works..
Its working, playing random files as it should. The .ino file is complete balingwire and bubble gum. Its a hacked mess. But whatever. The big issue is; I just updated all my Adafruit libraries and now?  It seems they are comflicting with each other. So that's it for tonight. It runs, after a fashon, and I don't dare try fixing anything.

-me!