IDnameBuff started. untested.
Not even sure this approach will work, but I have to try something. This is a base class that stores a buffer under a fileBlock ID  that has a name and a block of data. The plan is to referance it by the block ID and the name is for the user. to use as a label.