New Tac, just wrote contact code.
This is just the "It compiles but doesn't run yet" version. Writing filemaker was getting daunting. I couldn't wrap my brain around that big of a project all at once with the tools I had. So I downshifted and I'm starting smaller and less general puropse. Possibly doable. We'll see..