Breakout is up and running.
Good enough, its a pile of hack code to begin with.  Pretty neat I was able to port a hacked game that was never intended tor A touch control, B screen other than 128x128 and C sharing an OS and being able to be deleted and recalled.. In a day! This is a good OS model!