Breakout is done for now.
Solved the time lag issue. Turs out if the panel "wants" clicks it'll cause a complete redraw as soon as one happens. With a bunch of bitmapped pictures.. This can take awhile. While the timers are all running.. Turned that off and its silky smooth!

Buttons on on address list, call address is hooked in.