Adding events, its just a MESS!
Adding in events either I have loose pointers or ran out of RAM but nothing works anymore and it is acting like a Stack/Heap collision. I need to get this into the server so I can head off to the shop to work on it. I need to get out of the house and away from Factorio.