Scrolling works, there are hacks though.
editField NEEDS a rewrite! It needs to have a decent rectangle, colors, scolling for strings too long to show when editing. All this stuff. The entire bailing wire thing where its part of another object is a gigantic kludge and only exisits to cover some if its design flaws.

Also, dragging needs to start not only when sitting too long, but when the finger starts moving. This it what people will expect.