Start working on some contact cleanup and why can't I make calls?
I've not been able to make a call for some time. So I took off the home screen picutre and put the cell status back on there. I can see that the tow guys are talking and the phone is talking to the network.  But I don't think there's enough signal around here to make the poor thing work. Did make it format phone number strings though.