Added bits to status. Removed recieve buffer.
Something has been causing the Feather to randomly reboot. It looks like this is why my signal has been so bad. as soon as I get some, pretty quickly the poor thing will reboot and start all over again. I mad the input buffer static just in case it was a leak that I was just not seeing. I've not seen the issue since, but now I see that it's locking up when I try to make a call. Locking up in a handler. I also made the status code smarter in that when a command is dropped it cleats the status to an "unknown" state so you can see the damn thing has lost connection.