Update comments, made delay() reentrant safe.
Went through and added a BUNCH of coments to the library code for cell phone, block file and other things that were created for litlOS & the cell phone. Also. I finally added a flag to the idler code to tell when its in the idle loop or not. This way our delay() function wont cause a reentrant condition if called during the idle loop.