Still broken but I know more.
I saw bluggering come in. I was able to watch the stuff coming through the FONA port from the SIM and its actually a mess of junk with the stuff I want. There is extra from the clock string cloggingup things as well. I disabled the clock and that helped  lot.

I think I need to parse it out in the FONA. Sending over a ton a giberish seems to be a bad waste of bandwidh.