CallerID works.. Sometimes.
Found what was causing all the wierd errors. Errors in the GUI was making it look lke the FONA was crashing. Also I was looking for the caller ID in a buffer that was never filled up. So.. I hacked a call into the adafruit stuff to let it fill up between commands to see if there was a caller ID waiting to come in. Its starting to work! Even the lookup works. But it still misses the ID about 50% of the time. Have to work on that.