Moved checking caller ID to beginning.
Intead of waiting 'till there is no command coming in before checking caller ID, we just check it whenever we see a command on the way, but before we take action on that command. So we are insured on the fullest plupest, juciest buffers possible. (Not tested, tomorrow)

Also added a bit to the setting window so it only updates when it sees a change in the status number.