The file was missing?!
Opened up the Phone to find that the screen background image was actually missing. Replaced the file and everything was fine.

Now there's an issue with the  contacts. I can't seem to see any and when I tried to reset the contact file it locked up the machine. I'm thking I may have damaged the SD drive yesterday when I accadentlally dropped the phone yesterday when I was setting up to do the tests on it. I'll have to grab a new SD card and see if that'll solve the contact's issue.