Disk issues fixed.
First I think I broke the 2nd SD card. Adding back in the screen images fixed the screen the new issue was that I couldn't create the new default contact list without crashing. The deal? I was handing in what seems to be a c string constand to PNLabel's filterPNStr() and that writes to the incoming string. Pass in a string constant is instant death.

I ended up rewriting contactList::findByPN(char* phoneNum)

To use an internal buffer before handing off to filterPNStr(). To take care of 90+% of the issues and wrote a warning on filterPNStr()'s comments about it.