Broke up Flora II into objects.
The program was getting way out of hand so I broke it into..
parameters
pumpObj
textComObj

This simplified things a lot, although it was a little tough making the breaks. I Now want to break off some sort of com object for the handheld.

Also got manual PWM working. So the pump speed can be controlled. This is hopefully going to allow  variable flow rates while watering.