Tons of changes. BIG one is menuBar
panels now have menu bars biuilt in with defaut of working close boxes. ecept for your home screen because there's nowhere to go back to if you close it. Bunch of changes for flora stuff including FONTS and addition of a base bitmap object. I'mgoing to see if I can redice a bunch of code by basing bitmap stuff on it.