File transfer works.. Slowly.
It gets a 254 byte buffer every second or so. This'll have to be worked on. The bit that handles the comunicatoins only hits the wire every so often as set by a timer. I'll have to "spectial case" the file transfer stuff because its going to need to pump over a lot. the issue is that the UI and tthe file transfer bash heads for time.