Added color editor
First pass at color editing. This incliuded adding EEPROM support to save the edits along with going to a 24 hour clock format starting at 0.

Screen drawing is working, but its just  a hack. The fix is to go back to how it was supposed to be drawn in the first place. With objects as opposed to funky direct screen drawing.

liltParser NEEDS to get away from static buffer so we can dump in masses of date. Now it just crashes everything when it overruns oits buffer. Or at least fix te crashing bit.