Accepting calls  seem to be fine now.
There were issues with the phone thinking it didn't pick up a call when it actually did. I added a flash in the FONA code when it failed a pickup. Never saw the flash and never had another fialed pickup. I did up date the FONA lib files. Maybe that's what fixed it? Who knows?