Got evenLog running with example.
Wrote up eventLog base code and an example  .ino to test it. Also reverted the Adafruit 1947 code back to the adafruit drivers from the teensy ones. There was a lot of work debugging the SD drove and file stuff so I wanted it simplified. Hopefully I can change it back later to the faster stuff. (Wasn't the driver's fault)