Bunch of changes
Added breakout to the flora controller.

Added LC_piezoTunes for code to play piezo beepers, tunes and stuff. This was for playing sounds for the breakout game.

Moved sTerm to the scr folder in the floara controler.

Added a copy of AnimatedGIF library and got it working, just for fun.

Added alpha channel to my bitmap stuff and base displayObj. (Finally)