First app working WITHOUT back pointing crap!
Had to combine the panel files into the lilOS files and ditch the appKit file idea. But now we can access the main application's custom OS where all the configuration is done by the "standard" panels.  Yay! Although there is a LOT of work still to do. This at least proves the theory.