Color picker works.
Ended up doing a complete rewrite of slider library. I knew it wasn't used and it didn't work very well. There is an issue with the touchscreen giving bogus values on lift. So I filtered the out and, for now, that's fixing the issue. But, sadly, this will only work on a case by case deal. Need better fix.

Also found there is something going on messing up the paths possibly when deleting a folder? Not sure but the problem ALWAYS shows up on save where there is no files to see and no name on the folder.