Cleanup lilOS some. FIxed crash bug in grenade.
The gloabl pointer to the current OS object was fuzzy in the beginning of code. So I forced all the panels to use a version handed to them by the creating OS. No issues there. Then returned the global for different packages to access. And set it as the first item in the base OS's begin() call.

Making file path generation a base panel call broke some stuff. All fixed now. Took days to track down. Caused me to use a forgotten now uninitialized variable.