updated the RPN Calculator
Originally this was written as a stand alone app I tried to fit into an Arduino UNO. To attempt to save RAM, I used a lot of globals for colors, object addresses etc. This helped, but never was enough to shoe horn it into the UNO. Now, it's sharing RAM with other applications. In this case globals are a bad habit. So I did a giant update and got rid of all the globals so it plays nicer with it's neighbors. Makes more logical sense as well. Needed a rewrite.